Moved RouteError into error.rs
