+++ crates/web_ui/src/lib.rs
@@ -2,3 +2,4 @@ mod session;
 pub mod settings;
 pub mod ui;
 pub mod user;
+pub mod error;
